workers/server: AsyncOSBuildJobInfo
This commit brings in a new method into the server, AsyncOSBuildJobInfo.
That has for role to return a JobInfo and an OSBuildJobResult, same as
OSBuildJobInfo. The main difference relying on the fact that the
OSBuildJobResult is not fully complete.

After calling the AsyncOSBuildJobInfo function, the result contains only
this information:

* JobError
* Success
* OSBuildOutput.Success (if it has an OSBuildOutput)
* OSBuildOutput.Error (if it has an OSBuildOutput)
* TargetResults

Additionally, OSBuildOutput stays nil if no data associated to it is in
the DB.

Later on, the user can, if they want, make subsequent calls to add in
more information. By calling QueryResultFields(id, paths, any).

This function takes the id of the Result to get data from, the path in
the json result blob to fetch information from, and a pointer to a
JobResult to store this information in.