dbjobqueue: treat "null" answer as nil
Postgres will return "null" if the json contains this value for a given
field. This isn't an issue for unmarshalling as is will successfully
unpack that value to nil.

However, in a case where the field we want to recover from the DB is to
be stored under another one that happen to be also nil, the previous
implementation would've instantiated the parent level to allocate a nil
value to its field child (the one requested).

Then all kinds of bad behavior can happen, like if the user relies on
the absence of value of that parent field to take a decision for
instance.

The commit adds a new test that depicts a google example of what that
means.

The solution to that is to tread the literal "null" value as nil and to
avoid exploring the data structure, and thus avoiding creating instances
along the way.