image: add ostree commit source when initialising manifest
When initialising a manifest, use the ostree commit source spec only.
Then, when serialising, use the resolved ostree commit.

The fallback to the default ref when one isn't specified for images that
pull a commit is now handled in the image function.

This is the same process we use for the other content types, packages
and containers.

The meaning of the ostree commit content differs based on image type:
sometimes it serves as a parent for a newly built commit (commit and
container), while other times it is the payload content for a new image
(installers and raw-image).  This is reflected in the argument name of
the image function.