distro: allow empty ostree options in image functions
If an ostree ref is optional for an image type's image function (commits
and containers), don't rely on the options being set by the caller.  Use
the distro/image-type default if it's not set.

This puts the responsibility of setting ostree-based values on the image
type itself, rather than the caller.  In the same way that the image
function defines its own default content (i.e., packages), it also
defines its own ostree commit reference (returned as part of the
initialised manifest), unless it's overridden by a user option.

This better reflects the requirements for these image types: The ostree
commit ref isn't required, since the distro defines a default.  This is
not the same for image types that require a commit as payload
(installers and raw images), since those need a URL from which to pull
the content.