distro: use the ostree URL when checking options
The checksum won't be resolved before initialising the manifest (where
checkOptions() is called), so we can't rely on it for validating
options.  Instead, we can just make sure that the URL was provided, when
required by the image type.

Also, if a parent ref is specified in the options, a URL *must* be
specified regardless of image type, so verify this combination here
 and return the same error message that is returned from
 ostree.Resolve().

The parameter combination check will soon be removed from the
ostree.Resolve() function.