manifest: add package selector function to Manifest
Add a package selector function to the Manifest struct.

Packages are sometimes renamed between distribution versions and since
we do the package selection in the Manifest, we need a way to select
distro-version-specific package names inside the manifest initialiser.

For now, we add a package renaming function, but we call it
PackageSelector() because we intend to change it in the future to be a
distro-specific package selection function that relies on the image
configuration (Platform, Environment, Workload, etc) to decide which
packages should be added to each pipeline.