distro/fedora: derive ami from qcow2
Fedora doesn't build a separate AMI image. The ordinary Cloud Base image
is what's used for AWS.

Here's the code that's responsible for uploading ec2 images - it searches
for Cloud raw.xz images:

https://github.com/fedora-infra/fedimg/blob/fcbface13726cd1cb09b007b510bc888116a56e0/fedimg/consumers.py#L114

And here's the pungi configuration showing that qcow2 and raw.xz are actually
built from the same kickstart:

https://pagure.io/pungi-fedora/blob/e080c0702f38c033025889e5fcc8d9fee5bb2026/f/fedora-cloud.conf#_142

Thus, we can just do the same thing: Let's base the AMI on the qcow2 to save
some bytes.
Signed-off-by: Ondřej Budai <ondrej@budai.cz>