manifest: make ostree remote optional in raw images
In RHEL Edge Raw Image, we set the default ostree remote to be the
location from which we downloaded the ostree commit for the payload.
In Fedora IoT Raw Image, we would hardcode the default remote, but now
that Fedora packages the remote config, we want to remove the
configuration in the deployment.

Make the ostree remote in the pipeline optional so we can only set it
when it's needed.  The remote should be configured from the caller,
which is the image function in the distro.