blueprint: make Convert respect nils
Previously, nil values in the conversion source were in some cases converted
to empty arrays or empty objects. This is undesirable, because it can be in
certain cases changing the semantics of the blueprint. See e.g.
https://github.com/osbuild/images/blob/f317064da574a8475d438972e9c54bc67ffa1dfb/pkg/distro/rhel7/imagetype.go#L239C7-L239C7

This commit modifies the conversion process so nil values are converted
without any changes. Also, the `Convert` function was covered with a unit
test.