SPEC: enable default Go modules behavior on Fedora
It turns out, that Fedora Go RPM macros are disabling the Go modules
support by default for all builds. The reason seems to be the worry that
some Go macros may not support it. However, it is not disabled by
default on CentOS Stream or RHEL and the package builds just fine.

On Fedora, undefine the macro which explicitly passes `GO111MODULE=off`
to each build, to use the default compiler behavior instead.

Additional relevant info:
https://pagure.io/go-rpm-macros/c/35bd72648fe6cfa372f886b5692936ffb5b4389a
https://pagure.io/go-rpm-macros/c/df8452a5df762d8df1e871a103bc05a58c24cb8d

Signed-off-by: Tomáš Hozza <thozza@redhat.com>