Schutzfile: update osbuild SHA to v102
This should solve the CI failures on el8.10 due to fix the missing
`/etc/X11/xorg.conf.d/00-keyboard.conf` file when building images and
setting keyboard layout.

The osbuild v102 is currently blocked in RHEL CI, so workaround this
issue for now at least on `main` in the upstream.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>