aws: allow the user to customize their AMI name
In the frontend a new field is allowing user to setup customized names
for their images (see
https://github.com/RedHatInsights/image-builder-frontend/pull/1136).

To allow the customization to effectively take place, this commit is
slightly changing how EC2 images are activated.

The Name tag and the AMI Name are not always sharing the same value
exactly anymore. If the user provides a SnaphsotName, the raw value is
set in the `Name` tag and the AMI name is set to `$SnaphsotName-$UUID`
(to make it unique).

This means that the `Name` tag can't be used anymore for the maintenance
tooling. The tooling will have to use the new Build-By tag to identify
the images built by osbuild-composer.