Test/Koji: test only combinations that we run in the service
Test `koji.sh` with cloud upload on the latest RHEL-8.9 and RHEL9.3 GA
releases, which we use on our Brew workers.

Test the `koji.sh` only on the latest RHEL-8.9, RHEL-9.3 and Fedora
releases, which we use on our Brew and Koji workers.

There's no value in testing this case on CentOS Stream, RHEL EUS
releases or nightly composes.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>