packer: Fix installing worker RPMs in prod builds
We use --skip-tags rpmrepo_composer,rpmrepo_osbuild,rpmcopr for our prod
builds. Thus, if we merge the following two tasks into one, no RPMs will
be installed in prod builds, because --skip-tags joins multiple tags
with OR. (IoW a task with tags == [rpmcopy, rpmrepo_osbuild] will not
run with --skip-tags rpmrepo_osbuild).

This is quite a horrible hack, but it will hopefully unbreak us for the
time being.